+++ drivers/unix/dir_access_unix.h
@@ -80,7 +80,7 @@ public:
 	virtual Error rename(String p_path, String p_new_path);
 	virtual Error remove(String p_path);
 
-	virtual size_t get_space_left();
+	virtual int64_t get_space_left();
 
 	virtual String get_filesystem_type() const;
