+++ platform/android/dir_access_jandroid.h
@@ -78,7 +78,7 @@ public:
 	virtual String get_filesystem_type() const;
 
 	//virtual FileType get_file_type() const;
-	size_t get_space_left();
+	int64_t get_space_left();
 
 	static void setup(jobject p_io);
